+++ src/BUILD.plz
@@ -34,7 +34,6 @@ go_binary(
         "//src/test",
         "//src/tool",
         "//src/update",
-        "//src/utils",
         "//src/watch",
         "//src/worker",
         "//third_party/go:automaxprocs",
